stays tooltip completion
problem with tooltip: does not group several stays for the same
block/square. only uses the last one. need to find a way to describe
several stays and to represent that set.

maybe for a set of stays, use the highest value of time_spent of that
set

or try to sum all of the time_spent and use that value

and on the tooltip show the list of stays, much like the calendar